data handler extended for transaction

+++ src/asset.rs
@@ -1,16 +1,15 @@
 ///! Implementation of a container for basic asset data
+use serde::{Deserialize, Serialize};
 
-use serde::{Serialize,Deserialize};
-
-#[derive(Debug,Serialize,Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct AssetCategory {
-    id: u64,
+    id: usize,
     pub name: String,
 }
 
-#[derive(Debug,Serialize,Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct Asset {
-    pub id: Option<u64>,
+    pub id: Option<usize>,
     pub name: String,
     pub wkn: Option<String>,
     pub isin: Option<String>,
@@ -18,7 +17,19 @@ pub struct Asset {
 }
 
 impl Asset {
-    pub fn new(id: Option<u64>, name: &str, wkn: Option<String>, isin: Option<String>, note: Option<String>) -> Asset {
-        Asset{ id, name: name.to_string(), wkn, isin, note }
+    pub fn new(
+        id: Option<usize>,
+        name: &str,
+        wkn: Option<String>,
+        isin: Option<String>,
+        note: Option<String>,
+    ) -> Asset {
+        Asset {
+            id,
+            name: name.to_string(),
+            wkn,
+            isin,
+            note,
+        }
     }
 }
